// src/app/main/solicitudes-creditos/models/solicitudesCreditos.ts
export class SolicitudesCreditos {
    id: string;
    numero?: string;
    monto?: string;
    plazo?: string;
    aceptaTerminos?: string;
    estado?: string;
    user_id?: string;
    empresaComercial_id?: string;
    empresaIfis_id?: string;
    reporteBuro?: string;
    calificacionBuro?: string;
    buroValido?: string;
    identificacion?: string;
    ruc?: string;
    rolesPago?: string;
    panillaIESS?: string;
    created_at?: string;
    updated_at?: string;
    state?: string;

    tomarSolicitud?: string;
    tipoCredito?: string;
    concepto?: string;
    fechaAprobacion?: string;
    documentoAprobacion?: string;

}